requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.4.2
python-multipart==0.0.6
pyjwt==2.8.0 